NLP/notification-service/requirements.txt
sib-api-v3-sdk
python-dotenv
